net.conv: cast result of C.ntohs to return type

+++ vlib/net/conv/conv.c.v
@@ -2,12 +2,12 @@ module conv
 
 // host to net 32 (htonl)
 pub fn htn32(host &u32) u32 {
-	return C.htonl(host)
+	return u32(C.htonl(host))
 }
 
 // host to net 16 (htons)
 pub fn htn16(host &u16) u16 {
-	return C.htons(host)
+	return u16(C.htons(host))
 }
 
 // net to host 32 (ntohl)
@@ -17,5 +17,5 @@ pub fn nth32(host &u32) u32 {
 
 // net to host 16 (ntohs)
 pub fn nth16(host &u16) u16 {
-	return C.ntohs(host)
+	return u16(C.ntohs(host))
 }
